[#60]rename: stmt.model -> stmt.models
