+++ app/visualize/generator/converter/assign_converter.py
@@ -1,8 +1,8 @@
 from app.visualize.analysis.stmt.models.assign_stmt_obj import AssignStmtObj
 from app.visualize.generator.highlight.expr_highlight import ExprHighlight
 from app.visualize.generator.highlight.list_highlight import ListHighlight
-from app.visualize.generator.model.assign_viz import AssignViz
-from app.visualize.generator.model.variable_vlz import Variable
+from app.visualize.generator.models.assign_viz import AssignViz
+from app.visualize.generator.models.variable_vlz import Variable
 from app.visualize.generator.visualization_manager import VisualizationManager
 from app.visualize.utils.util import Util
