+++ app/visualize/generator/converter/expr_converter.py
@@ -1,7 +1,7 @@
 from app.visualize.analysis.stmt.models.expr_stmt_obj import ExprStmtObj
 from app.visualize.generator.highlight.expr_highlight import ExprHighlight
-from app.visualize.generator.model.expr_viz import ExprViz
-from app.visualize.generator.model.print_viz import PrintViz
+from app.visualize.generator.models.expr_viz import ExprViz
+from app.visualize.generator.models.print_viz import PrintViz
 from app.visualize.generator.highlight.list_highlight import ListHighlight
 from app.visualize.generator.visualization_manager import VisualizationManager
 from app.visualize.utils.util import Util
