+++ app/visualize/generator/converter/if_converter.py
@@ -4,7 +4,7 @@ from app.visualize.analysis.stmt.models.if_stmt_obj import (
     ElifConditionObj,
     ElseConditionObj,
 )
-from app.visualize.generator.model.if_viz import ConditionViz, IfElseDefineViz, IfElseChangeViz
+from app.visualize.generator.models.if_viz import ConditionViz, IfElseDefineViz, IfElseChangeViz
 from app.visualize.generator.visualization_manager import VisualizationManager
