+++ app/visualize/generator/models/print_viz.py
