+++ tests/visualize/generator/converter/test_assign_converter.py
@@ -5,8 +5,8 @@ import pytest
 from app.visualize.analysis.stmt.models.assign_stmt_obj import AssignStmtObj
 from app.visualize.analysis.stmt.models.expr_stmt_obj import ExprStmtObj
 from app.visualize.generator.converter.assign_converter import AssignConverter
-from app.visualize.generator.model.assign_viz import AssignViz
-from app.visualize.generator.model.variable_vlz import Variable
+from app.visualize.generator.models.assign_viz import AssignViz
+from app.visualize.generator.models.variable_vlz import Variable
 from app.visualize.generator.visualization_manager import VisualizationManager
