+++ tests/visualize/generator/converter/test_expr_converter.py
@@ -2,7 +2,7 @@ import pytest
 
 from app.visualize.analysis.stmt.models.expr_stmt_obj import ExprStmtObj
 from app.visualize.generator.converter.expr_converter import ExprConverter
-from app.visualize.generator.model.print_viz import PrintViz
+from app.visualize.generator.models.print_viz import PrintViz
 
 
 @pytest.fixture
